+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGES:
 * Slog provider moved to providers directory.
 * More strict check of vshard.storage.call response.
 * Bump go-tarantool from v2.3.0 to v2.3.1.
+* Get rid of nameToReplicasetMutex, use atomic instead.
 
 ## v2.0.5